Update Docker configuration
Updated the Docker configuration to use PHP 7.3 and NGINX to use the 
stable container. The change to PHP was mainly because 7.1’s outdated
now. The change to the NGINX Docker container was much more important, 
as it’s know to be a problem if the stable container’s not used.

For more information, refer to 
https://github.com/nginxinc/docker-nginx/issues/262.